Fix Azure-AKS deployment regression
Issue-ID: ACUMOS-3842

Create /var/log/acumos for docker-proxy if not mounted.
Save kube-config for default deployment domain.
Use Nexus internal service address if possible.
Use Acumos domain IP in LoadBalancer ingress setup.

+++ tools/setup_docker.sh
@@ -44,7 +44,9 @@ setup() {
     if [[ $(sudo apt-get purge -y docker-ce docker docker-engine docker.io) ]]; then
       echo "Purged docker-ce docker docker-engine docker.io"
     fi
-    sudo rm /etc/docker/daemon.json
+    if [[ -e /etc/docker/daemon.json ]]; then
+      sudo rm /etc/docker/daemon.json
+    fi
     wait_dpkg; sudo apt-get update
     wait_dpkg; sudo apt-get install -y \
       apt-transport-https \
